+++ src/irDevices/AirConditioner.ts
@@ -7,7 +7,7 @@ import {
 } from 'homebridge';
 import { SwitchBotPlatform } from '../platform';
 import { DeviceURL } from '../settings';
-import { irdevice, deviceStatusResponse } from '../configTypes';
+import { irdevice } from '../configTypes';
 
 /**
  * Platform Accessory
@@ -18,8 +18,6 @@ export class AirConditioner {
   service!: Service;
 
   Active!: CharacteristicValue;
-  ActiveIdentifier!: CharacteristicValue;
-  deviceStatus!: deviceStatusResponse;
   RotationSpeed!: number;
   lastTemperature!: number;
   currentTemperature!: number;
@@ -67,8 +65,6 @@ export class AirConditioner {
       .getCharacteristic(this.platform.Characteristic.Active)
       .on(CharacteristicEventTypes.SET, (value: any, callback: CharacteristicGetCallback) => {
         this.platform.log.debug('%s %s Set Active: %s', this.device.remoteType, this.accessory.displayName, value);
-        this.platform.log.warn(value);
-
         try {
           if (value === this.platform.Characteristic.Active.INACTIVE) {
             this.pushAirConditionerOffChanges();
@@ -130,8 +126,6 @@ export class AirConditioner {
       })
       .on(CharacteristicEventTypes.SET, (value: any, callback) => {
         this.platform.log.debug('');
-        this.platform.log.warn(value);
-
         if (value === 4) {
           this.currentFanSpeed = 1;
         } else {